Migrating Atlassian Plugins to Java 5, removing deprecated classes and methods PLUG-108
git-svn-id: https://studio.atlassian.com/svn/PLUG/trunk@22570 2c54a935-e501-0410-bc05-97a93f6bca70

+++ atlassian-plugins-core/src/main/java/com/atlassian/plugin/util/JavaVersionUtils.java
@@ -7,7 +7,7 @@ public class JavaVersionUtils
 {
     public static boolean satisfiesMinVersion(float versionNumber)
     {
-        float specVersion = Float.valueOf(System.getProperty("java.specification.version")).floatValue();
+        float specVersion = Float.valueOf(System.getProperty("java.specification.version"));
         return specVersion >= versionNumber;
     }
